Extract Assistance component
